+++ libs/query/src/lib/query/query.utils.ts
@@ -220,7 +220,7 @@ export const isQueryCollection = <T extends AnyQueryCollection>(query: unknown):
 };
 
 /**
- * @deprecated Use `createQueryCollectionSubject` instead. Will be removed in v5.
+ * @deprecated Use `createQueryCollectionSubject` instead. Will be removed in v6.
  */
 export const createQueryCollection = <T extends AnyQueryCreatorCollection, R extends QueryCollectionOf<T>>(
   // eslint-disable-next-line @typescript-eslint/no-unused-vars