+++ libs/types/src/lib/api/Round/index.ts
@@ -1,2 +1,2 @@
-export { RoundListView } from './roundList.view';
 export { RoundWithStageAndGroupView } from './roundWithStageAndGroup.view';
+export { RoundListView } from './roundList.view';