+++ libs/components/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ethlete/components
 
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ethlete/cdk@4.0.3
+
+## 3.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ethlete/cdk@4.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ethlete/cdk@4.0.1
+
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`2ded18ef`](https://github.com/ethlete-io/ethdk/commit/2ded18ef14115c9c9e2fb4f86c688d436c807766)]:
+  - @ethlete/cdk@4.0.0
+
 ## 2.6.1
 
 ### Patch Changes