+++ libs/contentful/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @ethlete/contentful
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ethlete/cdk@4.0.3
+  - @ethlete/core@4.0.3
+  - @ethlete/query@5.0.3
+
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ethlete/cdk@4.0.2
+  - @ethlete/core@4.0.2
+  - @ethlete/query@5.0.2
+
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`3f77e8d5`](https://github.com/ethlete-io/ethdk/commit/3f77e8d52a5ba45c4f3da4e34dcc08e0561ae04d)]:
+  - @ethlete/core@4.0.1
+  - @ethlete/cdk@4.0.1
+  - @ethlete/query@5.0.1
+
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`2ded18ef`](https://github.com/ethlete-io/ethdk/commit/2ded18ef14115c9c9e2fb4f86c688d436c807766)]:
+  - @ethlete/cdk@4.0.0
+  - @ethlete/core@4.0.0
+  - @ethlete/query@5.0.0
+
 ## 1.0.3
 
 ### Patch Changes